Implement authentication for api endpoints

+++ todoapp/routers/todos.py
@@ -5,6 +5,7 @@ from fastapi import APIRouter, Depends, HTTPException, Path
 import models
 from database import SessionLocal
 from sqlalchemy.orm import Session
+from .auth import get_current_user
 
 router= APIRouter()
 
@@ -18,7 +19,7 @@ def get_db():
         db.close()
 
 db_dependency=Annotated[Session,Depends(get_db)]
-
+user_dependency=Annotated[dict,Depends(get_current_user)]
 
 # For body structure and syntax validation
 class ToDoRequest(BaseModel):
@@ -28,28 +29,42 @@ class ToDoRequest(BaseModel):
     priority:int= Field(gt=0,lt=6)
     complete:bool
 
-
+# Returns all todos as per who is owner, using owner_id
 @router.get("/",status_code=status.HTTP_200_OK)
-async def read_all(db:db_dependency):
-    return db.query(models.Todos).all()
+async def read_all(user:user_dependency,db:db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    return db.query(models.Todos).filter(models.Todos.owner_id==user.get('id')).all()
 
 @router.get("/todo/{todo_id}",status_code=status.HTTP_200_OK)
-async def  read_todo(db:db_dependency,todo_id:int=Path(gt=0)):
-    todo_model= db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+async def  read_todo(user:user_dependency,db:db_dependency,todo_id:int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    
+    todo_model= db.query(models.Todos).filter(models.Todos.id == todo_id).filter(models.Todos.owner_id==user.get('id')).first()
     if todo_model is not None:
-        return todo_model
+        return todo_model             
     raise HTTPException(status_code=404,detail="Todo not found")
 
 @router.post("/todo",status_code=status.HTTP_201_CREATED)
-async def  create_todo(db:db_dependency,todo_request:ToDoRequest):
-    todo_model=models.Todos(**todo_request.model_dump())
+async def  create_todo(user:user_dependency, db:db_dependency,todo_request:ToDoRequest):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    
+    todo_model=models.Todos(**todo_request.model_dump(),owner_id=user.get('id'))
     db.add(todo_model)
     db.commit()
     # Finalize operation
 
 @router.put("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db:db_dependency,todo_request:ToDoRequest,todo_id:int=Path(gt=0)):
-     todo_model= db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+async def update_todo(user:user_dependency ,db:db_dependency,todo_request:ToDoRequest,todo_id:int=Path(gt=0)):
+     
+     if user is None:
+         raise HTTPException(status_code=401,detail="Authentication Failed")
+       # Authentication will fail here if details are incorrect  
+         
+     
+     todo_model= db.query(models.Todos).filter(models.Todos.id == todo_id).filter(models.Todos.owner_id==user.get('id')).first()
      if todo_model is None:
          raise HTTPException(status_code=404,detail="Todo not found")
      todo_model.title=todo_request.title
@@ -61,9 +76,12 @@ async def update_todo(db:db_dependency,todo_request:ToDoRequest,todo_id:int=Path
      db.commit()
      # Finalize operation.
 
+# Deletion function with user authentication
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency,todo_id:int=Path(gt=0)):
-    todo_model= db.query(models.Todos).filter(models.Todos.id == todo_id).first()
+async def delete_todo(user:user_dependency,db:db_dependency,todo_id:int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    todo_model= db.query(models.Todos).filter(models.Todos.id == todo_id).filter(models.Todos.owner_id==user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404,detail="Todo not found")
     db.delete(todo_model)
